test/pyproject.toml
[project]
name = "nike-crawler"
version = "0.1.0"
description = "네이버 쇼핑 나이키 신발 크롤링 애플리케이션"
authors = [
    {name = "Developer", email = "dev@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.40.0",
    "pandas>=2.1.4",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "pyyaml>=6.0.1",
    "aiohttp>=3.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/nike_crawler"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
